requirements.txt
python-telegram-bot==20.3
google-generativeai>=0.3.0
pymongo>=4.5.0
requests>=2.31.0
python-dotenv>=0.21.0
python-dateutil>=2.8.2